Feat : 로그인 ViewModel 구현 #13
- 기존 기능대로 동작, Firestore 유저 정보 저장 기능 미구현
